Move `StructureNameTable` to `StructureCatalog.cpp`
